Migrate empty disposable to DisposableHelper.
This also makes a hugely important change in that `Disposables.disposed()` is now synonymous with `Disposables.empty()`. This prevents `DisposableHelper.DISPOSED` from being leaked into the public API. If this were to be used by public code it could have have Very Bad™ effects by causing operators to think that they already disposed resources when they actually had not. Since `Disposable` is stateless it's safe for `disposed()` and `empty()` to be synonymous, although that might change in the future so both methods are retained.